Fix #1 - Sound was cutting out
There was an issue in our port from q3e/iortcw where it was still considering the loopSounds array as being an entityNumber and that was an issue. entityPositions were not updated either.

Also merged in some fixes from Q3e in the sound code that we didn't have yet. Seems to have helped with an extra crash that was happening too.