Steam Integration Only works Windows and Linux*
If com_steamIntegration is 1 on startup (maybe it can default to 1?), the game will look for a steam_api.dll and steam_appid.txt and attempt to initialize the Steam API. With an extra download of these two files, one can have the Steam Overlay and hours tracking without launching W:ET through Steam itself.

https://github.com/eternalcodes/EternalJK/commit/d28f4888929df5af63dcf662a6a25c908695c12c
Credits to @Bucky21659 from EternalJK implementation and @eezstreet for the original implementation inspired.

Overlay does not work on Linux at the moment. But steam detection does work with a correct steam api shared.