renderers: fix wrong decal projectors being culled
The wrong decal projectors may be culled if not all are visible and there are less than or equal to 32 decal projectors.

decalBits and numDecalProjectors were always set as if the active decal projectors were moved to the beginning of the list. However this only happened if there was more than 32 decal projectors.

The issue was introduced in commit https://github.com/etlegacy/etlegacy/commit/410d717e91ff7008430751cdce46842a66d3a841,
"renderers: make decals work with multiple world scenes".

Co-authored-by: @ryzyk-krzysiek
Refs https://github.com/etlegacy/etlegacy/issues/2072